pyproject.toml
[project]
name = "zerobus-databricks"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "databricks-sql-connector>=2.0.2",
    "faker>=40.1.2",
    "fastapi[standard]>=0.128.0",
    "pandas>=3.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
]
